ai-service/requirements.txt
fastapi[standard]
e2b_code_interpreter
pydantic-settings
loguru
langchain
langchain-openai
langchain-community
